Add modifier stacking rules
Add several classes for implementing the modifier stacking rules. This is in preparation for later use on the various actor statistics using the regular modifier stacking rules.

Use modifier stacking rules for PC saving throws by replacing the existing total modifier calculation with the new modifier stacking rules, which will allow a more expressive breakdown of the modifiers for each saving throw. This serves as an example of how to use the modifier stacking rules.