Unify apostrophes in item names.
A number of items had a name with a single quotation mark (’, &rsquo;)
rather than an apostrophe (', &apos;). This led to unintended behavior
when searching for items where seemingly identical items would not be
treated the same.

E.g. Searching for "Healer's" would find "Expanded Healer's Tools" but
not "Healer’s Tools".