Support for default roll options
Add support for default roll options, which is an list of flags that can contain predefined and arbitrary options that can be toggled to dynamically enable and disable modifiers on various checks and rolls. This mechanism already exists for damage rolls, and this change will add a similar mechanism for attack rolls, perception checks, initiative, saving throws, and skill checks.