Comment out Loot sheet edit checkbox
as was brought up in the foundry discord, the edit checkbox currently
does nothing except bring up a less useful version of the loot sheet.

Until something functional is driven by this edit box, it shouldn't be
there.

Note: The only thing that was only in the edit sheet was the name edit.
For now, it's only available to the GM.