Custom modifiers for NPC perception
Refactor the actor class to use custom modifiers for perception. The modifiers will currently consist of the ability modifier as well as a "base modifier", which is calculated from the freeform value of the NPC actor data.

Remove a previsously added tooltip breakdown from the editable NPC sheet, as the output included custom modifiers and would not match the base value presented in the input box in the UI. This could lead to confusion for the user.