Support for custom Dexterity modifier caps
Support for adding a list of arbitrary Dexterity modifier caps,of which the cap with the lowest value will be automatically applied. The Dex cap from any equipped armor is automatically considered and does not have to be added explicitly.

Expose a couple of methods on the actor to add and remove these custom Dex caps.