Tighten up the effects tab
Add a more dedicated style sheet for the effects tab instead of re-using the actions one. Include the new effects items in the effects tab (duh!). Prevent useless dice rolling for both conditions and effects, but instead dump a basic chat card with image, name, and description to the chat.