Target flat-footed checkbox on actions tab
Add a checkbox at the top of the actions tab in the character sheet which will allow you to toggle the 'target:flatFooted' roll option. This is useful for enabling modifiers or extra damage dice triggered by the flat-footed condition when the target is only flat-footed relative to this character.