Support damage type and category for flat modifiers
Add support for damageType and damageCategory fields in the flat modifier rule elements, and as parameters for the Actor#addCustomModifier method. The situation where the base damage type is the same but using a different damage category is also supported. This will enable implementation of the backstabber trait in a later change.