Mark effects as expired
Add a start time to effects when dragged to the token or sheet, some fields for registering duration in effect sheet, and calculate the expired status and any remaining duration in the effect panel. Expired effects are marked with a label in the effect panel UI to alert the player that it should be removed.