Refine identification data
Refine the data applied to an item on identification status change. You have the option of linking a specific identification status to another item, from the world or compendium, and to override specific fields. There is also a default naming scheme for unidentified items.

This mechanism should trigger also on `item.update(...)` calls, whether they come from the UI, a macro, or the system itself.