Refactor the migration runner code to be much easier to maintain.
Break the migrations into individual files that they only need to have concerns of updating an item or actor directly, and not think about making sure to go through every item on every actor on every token in every scene. The runner handles  async code correctly so there won't be any race conditions, and the object is in a consistent state.