Add effect target rule element
Add an effect target rule element that will support adding a dropdown to the effect item sheet for selecting an item to have its ID injected somewhere in the effect data for use by injection in other rule elements to target specific weapons etc. This is useful for spells like Magic Weapon.