Add localization of skill labels to getdata function of the simple-npc-sheet.
This is basically just a hack, because it does only change the displayed name. But the large problem is coming from the fact that the name is used for various functions in the simple npc sheet. I opted to changing only the label, as changing the name leads to ugly side effects I am not able and willing to fix.