Fix XP macro reward and rating values
The XP macro always reports an encounter's reward as the Moderate
encounter budget. Update `calculateXP()` to return the encounter's
reward value, so the macro can properly report it.

The XP macro reports the `xpPerPlayer` value for the encounter's rating
value. This value is confusing though as the encounter budget table that
is displayed is already adjusted per player.