Add creature table to XP macro
Add the creature table to the XP macro. With the Encounter Budget and
Creature XP tables, the macro can be used to balance encounters.

Also add a new row to the Encounter Budget table that shows the amount
of XP needed to reach each threat level. This can be used to rebalance
existing encounters. For example when adjusting an adventure path
4-player encounter to another party size.