Remember the active actions tab when drawing sheet
The active tab in the Actions menu always resets to the first tab
anytime the character sheet is redrawn. This makes using the Exploration
and Downtime tabs annoying, as adding or deleting actions will cause the
sheet to switch back to the Encounter tab.

Switch the action tabs to use FoundryVTT Tabs, so that the active tab is
remembered when drawing the screen.