Fix drag-and-drop strikes on NPC sheet
The pf2e system is defining a custom DragEvent type for "Actions", so
that it can generate fancy hotbar macros.

But overriding the base Foundry DragEvent type means those "Actions"
can't be Foundry "Items". So drag-and-drop won't work anywhere Foundry
expects an Item, for example dragging to the Items sidebar. This is fine
for the character sheet as the Strikes aren't real Items, but on the NPC
sheet they are.

Fix this by migrating PF2e's custom DragEvent types to be in a separate
location in the DragEvent. This way a DragEvent can have both a base
Foundry type and PF2e type.