Fix extractPacks breaking on non-json files
Some people make life difficult for themselves and use the Vim text
editor. By default Vim drops hidden swap files in the directory of a
file being edited. When extractPacks was scanning for jsons, it would
grab the swap files and try to parse them as jsons, which would break.