Add Death Note to ActionsPF2e
There are a fair number of monsters that have special effects when they
die. For example the Bloodsiphon explodes upon death. GMs sometimes
forget to resolve these actions as they are out of the normal flow of
a combat encounter.

Add a new setting to Actions that tells the system it should post the
Action to chat when the owning actor dies.