Add treat poison action macro
For the Treat Poison action, add action implementation to the system and expose it on the pf2e namespace of the game object, add an action macro to call that system implementation, and also add a link to call the action from the action item description. Also add three separate effects for the relevant degrees of success, and link those from the action result chat card.