Add sense direction action macro
For the sense direction exploration activity, add action implementation to the system and expose it on the pf2e namespace of the game object, add an action macro to call that system implementation, and also add a link to call the action from the feat item description. Also add a RollOption rule element to several items to indicate that they can serve as a compass.