Show "post to chat" on hover of inventory items
Add the fa-comment-alt icon to show up when hovering over inventory
items, to show that clicking the icon will post to chat. To that end,
update all .item-image to use same base style, to make things simpler.
This also converts a bunch of older .item-image to use the
fa-comment-alt icon instead of the d20 roll icon. But this is
appropriate as none of those actually do a roll.