Add basic support for auras that transmit effects to allies
An aura is defined as a rule element ("Aura") and can include a list of effects that should be transmitted to anyone inside it. An effect may be transmitted to allies, enemies, or everyone and may require a check. With this initial version of the feature, only allied effects that require no check are fully implemented.

A reference implementation is included consisting of the Marshal's Aura from the Marshal Dedication.