Refactor CompendiumBrowser listeners
This technically isn't a v10 change but I noticed the infinite scrolling in
the compendium browser result list was getting stuck after a re-render if
the mouse wasn't moved.
I've changed the scroll event to append the new results to the DOM instead
of re-rendering the whole application.
I've also removed JQuery from the other listeners while I was at it.