Refactor wands and scrolls to not require embedded spell data
Newly created wands and spells will only have a UUID and some information
about the embedded spell.
Old items with the embedded spell data are still fully supported but will
show a warning in the console that that will change in a future system
version.
Old chat messages with the spell data embedded in the DOM wil no longer
work.