Fix formatting for multiline tagify tags
Height is changed to min-height to allow for expanding with multiple rows,
like the other tag height used for Prerequisites.

The Select tag shouldn't expand though, and it tries to expand even with
a single row, so its height is set separately.

Tagify's default shadow isn't thick enough to cover multiple lines - it
looks like an underline for 2 lines of text, and a block for more.
Expanding the shadow to 10em should prevent that for up to 20-line tags
which is ridiculously long, and doesn't seem to have any negative side effects.