Fix item:damage-dice:N roll options being set for the number of faces
There seems to have been a minor copy/paste error, causing this roll
option to be set for the number of faces of the damage die rather than
the number of dice. This causes an issue with the rule elements in the
Student of the Staff feat, causing the arcane cascade critical damage to
be rolled incorrectly.