Allow injected property values in 'option' property
Currently the RollOption RuleElement does not function correctly as a toggleable if the "option" property has been interpolated with an injected property. 

This is because when the `toggleOption` method is called, it iterates the RuleElements to find the matching RuleElement that this toggle was created from. When doing this comparison, the toggle object uses the already-injected version of "option", but it compares to the unmodified version of the RuleElement, meaning that it will never find the correct RuleElement.

To fix this, we can call "resolveInjectedProperties" on the RuleElements' option properties as we iterate and compare, so that the Toggle `option` will indeed be a match for the interpolated RuleElement `option`.