Update roll-option.ts
Sorry I totally missed that the toggleOption method was `static`.

This new solution is a little more roundabout, since (to the level of my understanding) we need to reconstruct the RollOptionRuleElements to be able to accurately resolve the option parameter.

Although this does impact performance, the impact should be minor, since this method is only called upon user click, and it will only reconstruct those RollOptionRuleElements on the same item that could be a match for this toggle action.

If there's a simpler way to trigger the option resolution object that I'm not aware of let me know.