Add EphemeralEffect rule element
This rule element allows an attacker to push an effect onto a target for a single attack or damage roll--and also for the target to push an effect onto the attacker for the same roll. This is done by adding effect items to contextual clones while assembling attack/damage roll contexts.