Refactor action macro internal API (#6642)
Change the internal action check implementation to allow more flexible decision about which statistic to use when rolling the check. This is needed for actions where multiple statistics can be applied, like the Escape action. Provide a method to build the default check context that can bbe used in the vast majority of actions where only a single statistic is considered to ease the burden of implementing these actions.