Make it possible to suppress or override an effect applied on action use
Add a property to the action use options that allows the caller to suppress (value of `false`) or override (a UUID value or Effect item instance) the effect applied to the actor when the action is used. This is needed for specialized cases of certain action, like Take Cover with a tower shield that might not necessarily grant the usual effect of the action.