Implement shorthand syntax for multiple options in a rollable check
This allows using syntax like @Check[type:athletics,acrobatics] to
create two rollable buttons at once. When this syntax is used, there are
a few changes:

- The rollable buttons are created in a new block of text (div) rather
  than inline, with each button occupying a single line.
- The rollable buttons are "grouped", and posting one to chat will post
  all of the buttons to chat.
- The post to chat button uses a different icon (fa-comment-alt-dots) to
  indicate the buttons are grouped.

Otherwise, everything else works the same. Parameters and display text
are applied to all rollable buttons. In the future, I'm planning to
expand this functionality to allow the `adjustment` parameter to be used
to tweak the DCs of the different skills.