Adjust token-visibility testing to be more compliant with 2e rules
A further improvement would be to make the LoS polygon a composite of circles at multiple origin points,
but it'll be harder to get to the client code that creates the polygon.

This isn't in reference to an explicit vision rule but related to a diagram explaining cover rules: it is assumed that
Kyra _can_ (barely) see the ogre.