Update openSpellTab and compendium for remaster
FocusPool opens with Focus trait and I made up a filter called "Type" (not the
best name but it is 2am) to be able to select if you want to see rituals or
not. Then I made it where if you open the spell browser from the ritual entry
then the ritual filter is pre-selected.

The new options need to be moved to the proper place in config, and I don't
think we want `focus` and `ritual` hardcoded in openSpellBook but I'm out of
time for now.