Add very rudimentary action browser application
Add an application that lists actions sourced from the built-in system actions. The actions are split into tabs based on a section field in the action data, defaulting to the "other" tab. The built-in list of actions can potentially be expanded by modules.

Disable the use and check buttons in case no actor is controlled by, or assigned to, the user. In case of only a single controlled or assigned actor, show the relevant modifier on any check buttons.

Also add a control tool for opening the action browser (only in development mode for now).