Automate more of the Robe of the Archmagi effects
This automates most of the effects of a non-arcane caster wearing a Robe
of the Archmagi (regular or plus-sized).

The AC and the resilient item bonuses are supressed.  They'll appear on
the character sheet with the normal +2 value, but be grayed out and not
applied.

The good/neutral/evil alignment of the robe, and the robe's color, isn't
part of the item.  So none of the effects consider that.