Add rule for Divine Disharmony bonus
Rule for the +2 circumstance bonus to the divine disharmony action, when
the target meets certain conditions.

While the text of the feat mentions GM arbitration in determining the
nature of the target's faith or lack thereof, it also lists several
specific traits.  The RE tests only for the extact traits listed in the
text.