Set slug for resilient rune modifier (#13394)
When the modifier for a resilient rune is created, the adjustments
weren't also extracted and added to it.  The result is that
AdjustModifier REs can't change the resilient bonus.

This extracts the adjustments, in the same way as, e.g., the Bulwark
trait.  Items like Robe of the Archmagi that only supply their resilient
benefit if the owner has certain traits can be automated using an
AdjustModifier RE.