Automate Keen Follower feat
This will add the larger bonus when following an Expert or Master.

To get the necessary information, the Follow The Expert effect makes the
proficiency ChoiceSet copy to actor flags and a roll option.

The proficiency roll option is used so the Keen Follower modifier can be
disabled when it shouldn't apply, i.e. when following a trained or
legendary "expert".

The skill can't be used as a selector, since it might not be defined,
and this is a validation error.  The predicate would be false in this
case and the RE not used, but unapplied REs are still validated.

So skill-check is used as the selector.  But only the skill selected for
FTE will have the modifier on it to be adjusted, so all checks are not
modified.