Hide ordered list number in attack-popout
The attack list shouldn't show the "1.", "2.", etc. in the <OL> of
strikes.  The character sheet has this set to none for ol in general, but
the attack-popout doesn't match the same css selector.

While the margin settings mostly cut it off, you can see a bit of the digit
poking out on the edge of the window.