Type for Macro.execute() can accept additional properties for scope (#15086)
While actor and token are the only ones listed explicitly in the Macro.execute()
docs, the code can accept any additional properties and they will be placed into
the scope the macro executes in.  There's code in executeScript() that serves no
purpose other than allow this to be possible, so it's clearly intended.

This can be used to pass an event to the macro other than the global one or to
pass arbitrary parameters.