Refactor action outcomes to be separate from description and notes
Move the outcome from being part of the description and some roll notes to having its own structured data keyed off of degree of success.

Should help reduce duplicated text in translations (description and notes) and get rid of some of the embedded formatting, which should ease the burden for translators going forward.