Supply all roll callback parameters on attack with consumeAmmo
The consumeAmmo logic wraps an existing caller supplied attack roll
callback.  Unlike a RollParameter callback, which has only the Roll as
an argument, an attack roll callback is a CheckRollCallback, which has
four parameters.  The consumeAmmo wrapper would only pass along the
first of these.

Since the wrapper doesn't care about any of the arguments, use the rest
argument syntax to pass along whatever argument(s) the callback was
called with.  This way the code can work with any callback.

The callback also can be async, so await it, otherwise it won't finish before
the code in CheckPF2e.roll() that awaits the callback continues.