Update 4th-level tables' formatting
After running npm run extractPacks rollable-tables the 4th-level tables are modified despite not having been edited. Prior to running the command, each entry in the table has an empty flags key and the keys are in a different order. The items themselves are the same before and after tunning the command.