Reorder skills for treat wounds
Because skills that may be used instead of Medicine comes from a chosen feat, it is more likely that the non-medicine skill is preferred, so they are moved above Medicine in the "Treat Wounds Skill" drop-down. If the player does not have either of Chirurgeon or Natural Medicine, the Medicine skill will be showed on top.