Provide a subtitle with actions when using lores (#17223)
Current it will show a subtitle like "Lie (PF2E.ActionsCheck.legal-lore)" when
trying to use a skill that doesn't have a translation, which no lore will.

Detect the lack of translation and fall back to the "PF2E.ActionsCheck.x"
translation format using the label of the skill.  This will result in "Lie
(Legal Lore Check)" if the skill was labeled in that way.