Add applyOnce field to Weakness RE
Currently non-damage weaknesses are limited to a fixed set hard-coded in the
system with NON_DAMAGE_WEAKNESSES.

Add an optional field so weaknesses can be set to apply once.  If the field
is omitted, it will default to true if the weakness type was in
NON_DAMAGE_WEAKNESSES, false otherwise.

"non-damage" is renamed to "apply once".

Address issue #16190

This allows creating custom non-damage weaknesses.  E.g., a definition of
"origin:signature:{item|origin.signature}" would create a weakness to strikes
from a specific player who is the origin of the effect with the weakness.  By
setting applyOnce to true, the weakness will only trigger once for a damage
roll, rather than triggering on every damage instance in the roll.

It also makes it possible to create weakness to spells/items with a trait or
traits that do not trigger multiple times on spells with multiple damage types.
This already exists for the magic traditions and a few other traits, which are
in the NON_DAMAGE_WEAKNESSES set, but not arbitrary traits.  E.g., there's no
way to make a weakness to fear spells that does not trigger twice on damage from
Invoke Spirits, which deals both mental and void damage.