Add roll options for SingleCheckAction previews
When it generates the previews is only uses the roll option(s) that are
specific to the action, e.g. only "action:escape".  This doesn't match what
will be used for the actual roll.  For instance, it won't include
"armor:strength-requirement-met" and that results in the value being
incorrect for nearly every str or dex based action.