Add support for adding skill choices to Treat Wounds
Creates an array in flags to hold alternate skills to use for treat wounds and changes the treat wounds macro to add skills from there to its menu. This will generalize for any possible future features and homebrew additions.

May need explicit support for Chirurgeon and Natural Medicine added back to the macro unless existing features could be migrated to add the rule element to them.