Add roll data to consumable damage rolls
When rolling a consumable's damage roll, e.g. a healing potion, the damage roll
isn't supplied the roll data.  Add that.

This way something like `@actor.abilities.con.mod` can be used in the damage
roll.  This already works in an inline roll in the item's description, as that
is supplied the data.